[API-94] Split api/scripts/seed.ts into per-table inserters under scripts/seed/.
