+++ api/scripts/seed/.test.ts
@@ -1,9 +1,9 @@
 import path from 'node:path';
 import { describe, it, expect } from 'bun:test';
 import { schedules, zones } from '@/db/schema';
-import { computeInitialDepletionMm, seed, type SeedDb } from './seed';
+import { seed, type SeedDb } from '.';
 
-const SEEDS_DIR = path.resolve(import.meta.dir, '../data/seeds');
+const SEEDS_DIR = path.resolve(import.meta.dir, '../../data/seeds');
 
 function makeSeedDb() {
     const zoneConflictSets: Array<Record<string, unknown>> = [];
@@ -37,33 +37,6 @@ function makeSeedDb() {
 
     return { db, zoneConflictSets, zoneInsertValues, scheduleConflictSets, scheduleInsertValues };
 }
-
-describe('computeInitialDepletionMm', () => {
-    it('returns MAD when currentDepletionMm is not set', () => {
-        // MAD = 0.3 * 165 * 0.5 = 24.75  (clay-loam AWC = 165 mm/m)
-        const result = computeInitialDepletionMm(
-            { rootDepthM: 0.3, allowableDepletionFraction: 0.5, currentDepletionMm: undefined },
-            165,
-        );
-        expect(result).toBeCloseTo(24.75);
-    });
-
-    it('returns 0 when currentDepletionMm is explicitly 0', () => {
-        const result = computeInitialDepletionMm(
-            { rootDepthM: 0.3, allowableDepletionFraction: 0.5, currentDepletionMm: 0 },
-            165,
-        );
-        expect(result).toBe(0);
-    });
-
-    it('returns the explicit value when currentDepletionMm is set to a positive number', () => {
-        const result = computeInitialDepletionMm(
-            { rootDepthM: 0.3, allowableDepletionFraction: 0.5, currentDepletionMm: 12.5 },
-            165,
-        );
-        expect(result).toBe(12.5);
-    });
-});
 
 describe('seed zone upsert', () => {
     it('seeds zones at MAD (not zero) when currentDepletionMm is absent from the JSON', async () => {
